requirements.txt
pyttsx3
pysrt
pillow
numpy
moviepy